Fixed mess with types conversion.
Removed vw_Randf1 define.
Corrected comments.
Cosmetic changes.

+++ src/core/system/time_threads.cpp
@@ -108,13 +108,16 @@ float vw_GetTimeThread(int TimeThread)
 	// time manipulations
 	if (TimeThreadsMap[TimeThread].Speed != 1.0f) {
 		// calculate time from "time point" (DiffGetTicks), when speed was changed last time, till now
-		float RealTimeThread = ((SDL_GetTicks() - TimeThreadsMap[TimeThread].DiffGetTicks) *
+		// FIXME static_cast, we could have an issue with float type value
+		float RealTimeThread = (static_cast<float>(SDL_GetTicks() - TimeThreadsMap[TimeThread].DiffGetTicks) *
 					TimeThreadsMap[TimeThread].Speed) / 1000.0f;
 		// add "previous time" from time buffer
 		return RealTimeThread + TimeThreadsMap[TimeThread].Buffer;
 	}
 
-	return TimeThreadsMap[TimeThread].Buffer + (SDL_GetTicks() - TimeThreadsMap[TimeThread].DiffGetTicks) / 1000.0f;
+	// FIXME static_cast, we could have an issue with float type value
+	return TimeThreadsMap[TimeThread].Buffer +
+			static_cast<float>(SDL_GetTicks() - TimeThreadsMap[TimeThread].DiffGetTicks) / 1000.0f;
 }
 
 /*
@@ -154,7 +157,8 @@ void vw_SetTimeThreadSpeed(int TimeThread, float NewSpeed)
 	}
 
 	// store "previous time" in the time buffer
-	TimeThreadsMap[TimeThread].Buffer += ((SDL_GetTicks() - TimeThreadsMap[TimeThread].DiffGetTicks) *
+	// FIXME static_cast, we could have an issue with float type value
+	TimeThreadsMap[TimeThread].Buffer += (static_cast<float>(SDL_GetTicks() - TimeThreadsMap[TimeThread].DiffGetTicks) *
 					      TimeThreadsMap[TimeThread].Speed) / 1000.0f;
 	// store "time point", when speed was changed
 	TimeThreadsMap[TimeThread].DiffGetTicks = SDL_GetTicks();
